+++ .pre-commit-config.yaml
@@ -1,33 +1,43 @@
 repos:
--   repo: https://github.com/pre-commit/pre-commit-hooks
+  - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.2.0
     hooks:
-    -   id: trailing-whitespace
-    -   id: end-of-file-fixer
-    -   id: check-yaml
-    -   id: debug-statements
-    -   id: name-tests-test
-    -   id: requirements-txt-fixer
--   repo: https://github.com/asottile/setup-cfg-fmt
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
+      - id: check-yaml
+      - id: debug-statements
+      - id: name-tests-test
+      - id: requirements-txt-fixer
+      - id: check-ast
+      - id: check-json
+      - id: check-merge-conflict
+      - id: check-xml
+      - id: mixed-line-ending
+        args: ["--fix=auto"] # replace 'auto' with 'lf' to enforce Linux/Mac line endings or 'crlf' for Windows
+  - repo: https://github.com/asottile/setup-cfg-fmt
     rev: v1.20.1
     hooks:
-    -   id: setup-cfg-fmt
--   repo: https://github.com/janosh/format-ipy-cells
+      - id: setup-cfg-fmt
+  - repo: https://github.com/janosh/format-ipy-cells
     rev: v0.1.10
     hooks:
       - id: format-ipy-cells
-- repo: http://github.com/timothycrosley/isort
-  rev: 5.9.1
-  hooks:
-  - id: isort
-    args: ["--profile", "black", "--filter-files"]
-
-- repo: https://github.com/psf/black
-  rev: 22.3.0
-  hooks:
-  - id: black
--   repo: https://github.com/asottile/add-trailing-comma
+  - repo: http://github.com/timothycrosley/isort
+    rev: 5.9.1
+    hooks:
+      - id: isort
+        args: ["--profile", "black", "--filter-files"]
+  - repo: https://github.com/asottile/add-trailing-comma
     rev: v2.2.3
     hooks:
-    -   id: add-trailing-comma
+      - id: add-trailing-comma
         args: [--py36-plus]
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    rev: v0.0.288
+    hooks:
+      - id: ruff
+        args: [
+            "--line-length=320",
+            "--ignore=E203", # Ignore E203 specifically (known issue with Black)
+            "--fix", # Autofix any issues found by ruff
+          ]
