+++ AUTHORS.rst
@@ -1,0 +1,6 @@
+============
+Contributors
+============
+
+* Nabeel Seedat <seedatnabeel@gmail.com>
+* Nicolas Huynh <nvth2@cam.ac.uk>
