+++ requirements.txt
@@ -1,21 +1,21 @@
-synthcity==0.2.6
+synthcity==0.2.11
 langchain==0.0.228
 langchainplus-sdk==0.0.20
-matplotlib==3.7.2
-numpy==1.26.3
+matplotlib
+numpy<2.0
 openai==1.8.0
 safetensors==0.4.1
-scikit-learn==1.4.0
-scipy==1.11.4
+scikit-learn
+scipy
 sentencepiece==0.1.99
 seaborn==0.12.2
-vllm==0.2.7
+# vllm==0.2.7
 be_great==0.0.5
 imblearn==0.0
 openml==0.14.1
-pandas==1.4.3
-tqdm==4.64.0
-xgboost==1.7.6
-black==23.7.0
-isort==5.12.0
+pandas
+tqdm
+xgboost
+black
+isort
 umap-learn==0.5.3
