+++ scripts/great.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 
+# Get the directory of the current script
+SCRIPT_DIR=$(dirname "$(realpath "$0")")
+SCRIPT="$SCRIPT_DIR/../src/cllm/run_great.py"
+
 # Declare the arrays
 datasets=("covid" "seer" "compas" "cutract" "maggic")
 n_values=(10 50 100)
-seeds=(0 1 2)
+seeds=(0)
 
 # Iterate through the arrays
 for dataset in "${datasets[@]}"
@@ -13,8 +17,8 @@ do
     for runseed in "${seeds[@]}"
     do
       # Run the command
-      fuser -v /dev/nvidia0 -k
-      python ../src/run_great.py --dataset $dataset --runseed $runseed --ns $ns
+      # fuser -v /dev/nvidia0 -k # Use with caution
+      python "$SCRIPT" --dataset $dataset --seed $runseed --ns $ns
 
     done
   done
